Request to Promote "TEAMNADO-4774" to "prod-stable" (#190)
* TEAMNADO-4641: Activation Key Details - Save additional repos (#160)

* Add functionality to 'Add Repositories' table and model to allow repos to be added to activation key

* Fixed linting issues

* Fixed more linting errors

* Attempting to conform to linting rules

* Fixed final linting issues

* Made changes to resolve discussions

* Updated test cases to reflect previous changes

* Updated test

* TEAMNADO-4772 - Abstract common logic (#172)

* gated rbac button

* Fix linting errors

* Rename EditAdditionalRepositoriesButton to AddAdditionalRepositoriesButton

* update docs versions for Insights, Console, Cost & Subscriptions (#171)

* refactor addAdditionRepositoriesModal (#176)

* Unset repositories on toggle (#178)

* refactor addAdditionRepositoriesModal

* unset selectedRepositories on toggle

* "removed button wrapper, added error handle"

* "Testing changes, Comp name update, check added"

* "Components Name change"

* "/retest"

* "Removed error logic, attributes, testing ammends"

* "removed obsolete snapshot"

* "Repositories fix, null check removed"

* "key name added back accidental removal"

* "Component split, loading pattern amends, cap add"

---------
